* buffer.h, buffer.c (buf_output, buf_input_data, buf_read_line, buf_read_data, struct packetizing_buffer, packetizing_buffer_initialize) Use size_t instead of int. Silences warnings in buffer.c, server.c, and zlib.c on OpenBSD sparc64 where sizeof(int) is not the same as sizeof(size_t). * client.c (read_line_via, read_line, try_read_from_server, get_server_response, handle_ok, handle_error, handle_valid_requests, handle_checked_in, handle_new_entry, handle_checksum, handle_copy_file, handle_updated, handle_merged, handle_patched, handle_rcs_diff, handle_removed, handle_remove_entry, handle_set_static_directory, handle_clear_static_directory, handle_set_sticky, handle_clear_sticky, handle_clear_template, handle_module_expansion, handle_wrapper_rcs_option, handle_m, handle_e, handle_f, handle_notified): Ditto * client.h (struct response): Ditto. * server.c (receive_partial_file, receive_file, serve_modified, do_cvs_command): Ditto. * zlib.c (compress_buffer_input, compress_buffer_shutdown_input): Ditto.
CVS-ID: src/ChangeLog 1.2961
CVS-ID: src/buffer.c 1.50
CVS-ID: src/buffer.h 1.20
CVS-ID: src/client.c 1.398
CVS-ID: src/client.h 1.55
CVS-ID: src/server.c 1.389
CVS-ID: src/zlib.c 1.24

+++ src/buffer.h
@@ -97,7 +97,7 @@ struct buffer_data
     char *bufp;
 
     /* The number of data bytes found at BUFP.  */
-    int size;
+    size_t size;
 
     /*
      * Actual buffer.  This never changes after the structure is
@@ -127,12 +127,12 @@ struct buffer *buf_nonio_initialize (void (*) (struct buffer *));
 struct buffer *compress_buffer_initialize (struct buffer *, int, int,
 					   void (*) (struct buffer *));
 struct buffer *packetizing_buffer_initialize
-	(struct buffer *, int (*) (void *, const char *, char *, int),
-	 int (*) (void *, const char *, char *, int, int *), void *,
+	(struct buffer *, int (*) (void *, const char *, char *, size_t),
+	 int (*) (void *, const char *, char *, size_t, size_t *), void *,
 	 void (*) (struct buffer *));
 int buf_empty (struct buffer *);
 int buf_empty_p (struct buffer *);
-void buf_output (struct buffer *, const char *, int);
+void buf_output (struct buffer *, const char *, size_t);
 void buf_output0 (struct buffer *, const char *);
 void buf_append_char (struct buffer *, int);
 int buf_send_output (struct buffer *);
@@ -147,11 +147,11 @@ void buf_append_buffer (struct buffer *, struct buffer *);
 int buf_read_file (FILE *, long, struct buffer_data **, struct buffer_data **);
 int buf_read_file_to_eof (FILE *, struct buffer_data **,
 			  struct buffer_data **);
-int buf_input_data (struct buffer *, int *);
-int buf_read_line (struct buffer *, char **, int *);
+int buf_input_data (struct buffer *, size_t *);
+int buf_read_line (struct buffer *, char **, size_t *);
 int buf_read_short_line (struct buffer *buf, char **line, size_t *lenp,
                          size_t max);
-int buf_read_data (struct buffer *, int, char **, int *);
+int buf_read_data (struct buffer *, size_t, char **, size_t *);
 void buf_copy_lines (struct buffer *, struct buffer *, int);
 int buf_copy_counted (struct buffer *, struct buffer *, int *);
 int buf_chain_length (struct buffer_data *);
